requirements.txt
numpy>=1.24
pandas>=2.0
scikit-learn>=1.3
pyyaml>=6.0
matplotlib>=3.7
sentence-transformers>=3.0
transformers>=4.44
accelerate>=0.33
tqdm>=4.66
